refactor(paymasters): rename Peanut→Envelope, drop token allowlist, add per-tx ETH cap
The per-token allowlist was operator-side ceremony with little marginal safety:
the operator already curates which tokens get grants (by deciding what tx the
backend builds in step 2 of Path C). Removing it cuts an admin workflow.

Replaced with a per-tx ETH cap (`maxEthPerTx`, immutable, constructor-set) so
the worst-case drain under operator-key compromise is bounded per tx, not per
token. Combined with the existing daily QuotaControl cap, the security envelope
is equivalent for honest operation, tighter under compromise.

Renames (paymaster surface only; the vault keeps the upstream PeanutV4 name):
  - PeanutApprovalPaymaster        → EnvelopeApprovalPaymaster
  - peanutVault state              → envelopeVault
  - SpenderNotPeanut error         → SpenderNotEnvelope
  - EIP-712 domain name string     → "EnvelopeApprovalPaymaster"
  - GRANT_TYPEHASH                 → keccak256("EnvelopeApprovalGrant(...)")
  - file + test + deploy script names
  - all NatSpec and comments

NOTE: changing the EIP-712 domain name invalidates the signatures that would
verify against the previously-deployed paymaster at 0x301D...b6a0. That
contract is functionally orphaned now — needs a redeploy of the new bytecode
to a fresh address.

Tests: 19/19 envelope-paymaster (covers per-tx-cap, exceeded-quota via
constructor-tightened paymaster instance, sponsorship works on any token,
all the per-gate reverts). Full repo: 939/939, no regressions.
